requirements_hf.txt
anthropic>=0.40.0
gradio>=5.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
joblib>=1.3.2
scikit-learn>=1.4.0
numpy>=1.26.0
torch>=2.0.0
transformers>=4.30.0
